Remove ref to FMLDeobfuscatingRemapper in ObfuscationAwareAdapter
Accessing the map is completely unnecessary, since the transformer runs
post-deobfuscation (in the SRG namespace), so mapping names to Notch names makes
no sense.

The old code effectively caused idMethodName to get returned for methods in
Minecraft classes, and an empty string for modded classes. This commit emulates
this weird behavior to not risk breaking any code that relies on it.

Fixes incompatibility with CoreTweaks's transformer cache, which causes the map
to not get populated (this manifested as an issue with EFR's enchanting table.)